requirements.txt
feedparser
arxiv
newspaper4k
requests
ollama
